break out publish question
