+++ packages/geoprocessing/scripts/dataPrep/reimportData.ts
@@ -1,10 +1,8 @@
-import inquirer from "inquirer";
 import { reimportDatasources } from "../base/datasources";
 import { getProjectClient } from "../base/project/projectClient";
+import { publishQuestion } from "./publishQuestion";
 
-export interface PublishAnswers {
-  publish: "yes" | "no";
-}
+// This is a standalone script used as a CLI command with a top-level function
 
 const projectPath = process.argv[2];
 const matcher = process.argv[3];
@@ -19,26 +17,3 @@ void (async function () {
     doPublish: publishAnswers.publish === "yes" ? true : false,
   });
 })();
-
-export async function publishQuestion(): Promise<
-  Pick<PublishAnswers, "publish">
-> {
-  return inquirer.prompt<Pick<PublishAnswers, "publish">>([
-    {
-      type: "list",
-      name: "publish",
-      message: "Do you want to publish to S3 cloud storage now?",
-      default: "no",
-      choices: [
-        {
-          value: "yes",
-          name: "Yes",
-        },
-        {
-          value: "no",
-          name: "No",
-        },
-      ],
-    },
-  ]);
-}
